+++ src/Dto/Filters/AbstractFilter.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace OneSignal\Dto\Filters;
+
+use OneSignal\Dto\AbstractDto;
+
+abstract class AbstractFilter implements AbstractDto
+{
+}
