+++ src/Dto/Filters/BoughtSkuFilter.php
@@ -4,9 +4,7 @@ declare(strict_types=1);
 
 namespace OneSignal\Dto\Filters;
 
-use OneSignal\Dto\AbstractDto;
-
-class FilterBoughtSku implements AbstractDto
+class BoughtSkuFilter extends AbstractFilter
 {
     public const FIELD = 'bought_sku';
